feat: remove unused service
