+++ apps/mobile/android/app/build.gradle
@@ -1,8 +1,8 @@
 apply plugin: "com.android.application"
 apply plugin: "org.jetbrains.kotlin.android"
 apply plugin: "org.jetbrains.kotlin.kapt"
-apply plugin: "com.google.gms.google-services"
-apply plugin: 'com.google.firebase.crashlytics'
+// apply plugin: "com.google.gms.google-services"
+// apply plugin: 'com.google.firebase.crashlytics'
 apply plugin: "com.facebook.react"
 
 /**
@@ -120,9 +120,9 @@ android {
 dependencies {
     // The version of react-native is set by the React Native Gradle Plugin
     implementation("com.facebook.react:react-android")
-    implementation(platform("com.google.firebase:firebase-bom:32.7.1"))
-    implementation 'com.google.firebase:firebase-core:21.0.0'
-    implementation 'com.google.firebase:firebase-messaging:23.0.6'
+//    implementation(platform("com.google.firebase:firebase-bom:32.7.1"))
+//    implementation 'com.google.firebase:firebase-core:21.0.0'
+//    implementation 'com.google.firebase:firebase-messaging:23.0.6'
 
     if (hermesEnabled.toBoolean()) {
         implementation("com.facebook.react:hermes-android")
