+++ apps/mobile/android/build.gradle
@@ -18,8 +18,8 @@ buildscript {
         classpath("com.android.tools.build:gradle")
         classpath("com.facebook.react:react-native-gradle-plugin")
         classpath("org.jetbrains.kotlin:kotlin-gradle-plugin")
-        classpath 'com.google.gms:google-services:4.4.2'
-        classpath 'com.google.firebase:firebase-crashlytics-gradle:3.0.2'
+        // classpath 'com.google.gms:google-services:4.4.2'
+        // classpath 'com.google.firebase:firebase-crashlytics-gradle:3.0.2'
     }
 }
