+++ apps/mobile/ios/EurekaKit/EurekaKit.entitlements
@@ -4,11 +4,6 @@
 <dict>
 	<key>aps-environment</key>
 	<string>development</string>
-	<key>com.apple.developer.associated-domains</key>
-	<array>
-		<string>applinks:io.lemoncloud.eurekakit.app.goo.gl</string>
-		<string>applinks:eurekakit.page.link</string>
-	</array>
 	<key>com.apple.developer.applesignin</key>
 	<array>
 		<string>Default</string>
