--- apps/mobile/ios/GoogleService-Info.plist
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-    <key>CLIENT_ID</key>
-    <string>1049222889151-tip0e04mq7ceia1pjgla475i3jhphkqg.apps.googleusercontent.com</string>
-    <key>REVERSED_CLIENT_ID</key>
-    <string>com.googleusercontent.apps.1049222889151-tip0e04mq7ceia1pjgla475i3jhphkqg</string>
-    <key>API_KEY</key>
-    <string>AIzaSyCCrnpKhaDgGY3hpCNYnSJ66T4mhFK3ezA</string>
-    <key>GCM_SENDER_ID</key>
-    <string>1049222889151</string>
-    <key>PLIST_VERSION</key>
-    <string>1</string>
-    <key>BUNDLE_ID</key>
-    <string>io.lemoncloud.eurekapage</string>
-    <key>PROJECT_ID</key>
-    <string>eurekapage-85049</string>
-    <key>STORAGE_BUCKET</key>
-    <string>eurekapage-85049.firebasestorage.app</string>
-    <key>IS_ADS_ENABLED</key>
-    <false></false>
-    <key>IS_ANALYTICS_ENABLED</key>
-    <false></false>
-    <key>IS_APPINVITE_ENABLED</key>
-    <true></true>
-    <key>IS_GCM_ENABLED</key>
-    <true></true>
-    <key>IS_SIGNIN_ENABLED</key>
-    <true></true>
-    <key>GOOGLE_APP_ID</key>
-    <string>1:1049222889151:ios:2733438c784c161c44100a</string>
-</dict>
-</plist>
