+++ package.json
@@ -62,11 +62,6 @@
         "@react-native-camera-roll/camera-roll": "^7.10.0",
         "@react-native-community/netinfo": "^11.4.1",
         "@react-native-community/push-notification-ios": "^1.11.0",
-        "@react-native-firebase/app": "^20.3.0",
-        "@react-native-firebase/crashlytics": "^20.3.0",
-        "@react-native-firebase/dynamic-links": "^20.3.0",
-        "@react-native-firebase/installations": "^20.3.0",
-        "@react-native-firebase/messaging": "^20.3.0",
         "@react-native-google-signin/google-signin": "^12.2.1",
         "@react-native-seoul/kakao-login": "^5.4.1",
         "@react-native-seoul/naver-login": "^4.0.1",
@@ -103,7 +98,6 @@
         "react-native-keyboard-aware-scroll-view": "^0.9.5",
         "react-native-localize": "^3.4.1",
         "react-native-modal": "^14.0.0-rc.1",
-        "react-native-push-notification": "^8.1.1",
         "react-native-reanimated": "^3.17.3",
         "react-native-restart": "^0.0.27",
         "react-native-safe-area-context": "^5.3.0",
